New reader-based journal framework
This commit changes the journal architecture from being an in-memory
datastructure to being an abstraction over a io::Read like concept.

With this diff in place, it's possible to write new Journal
implementations backed by different storage layers. This is a key
requirement for adding persistence to sqlsync.

An in-memory journal has been implemented as an example.